Fixed $http error handling showing up blank
There is a right way to add commit messages and then there is my way. First, I need to remember
to commit for every logical working change before I find myself in another mess I can't get out from.
However, in this case it was good I couldn't because then I might have missed the reason it
wasn't working. After struggling for a while, and as a side and personal note, I have to remember
to use developer tools. The error I was getting was due to a dependency being misspelled.